requirements.txt
opencv==3.4.2
torch==1.10.2
torchvision==0.11.3
pillow==9.0.1
numpy==1.21.5
streamlit==1.11.0
